% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAD.R
\name{EAD}
\alias{EAD}
\title{Equivalent Air Depth (EAD) Calculation for Nitrox}
\usage{
EAD(EANx = 21, depth = NULL)
}
\arguments{
\item{EANx}{Percentage of oxygen in the EANx mix (default is 21\%).}

\item{depth}{Depth in metres.}
}
\value{
A character vector indicating the EAD.
}
\description{
Function to calculate Equivalent Air Depth (EAD) for EANx.
}
\examples{

EAD(EANx=21,depth=11)
EAD(27, 9)

}
