% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_RF.R
\name{get_RF}
\alias{get_RF}
\title{Function to calculate Repetitive Factors (RF) for DCIEM tables}
\usage{
get_RF(RG = NULL, surfaceinterval = NULL)
}
\arguments{
\item{RG}{Repetitive Group}

\item{surfaceinterval}{Surface Interval (minutes)}
}
\value{
A numeric vector.
}
\description{
Function to calculate Repetitive Factors (RF) for DCIEM tables
}
\examples{

get_RF(RG = "A", surfaceinterval = 50)
get_RF("C", 16)

}
