% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_RG.R
\name{get_RG}
\alias{get_RG}
\title{Calculate Repetitive Groups (RG) for DCIEM Tables}
\usage{
get_RG(depth = NULL, bottomtime = NULL, RF = 1, EANx = 21, ...)
}
\arguments{
\item{depth}{Depth in metres.}

\item{bottomtime}{Bottom time in minutes.}

\item{RF}{Repetitive factor.}

\item{EANx}{Enriched air nitrox (EANx) mix.}
}
\value{
A character vector indicating the Repetitive Group.
}
\description{
Function to calculate the Repetitive Group (RG) based on DCIEM diving tables.
}
\examples{

get_RG(depth = 5, bottomtime = 50)
get_RG(19, 30)

}
